+++ tests/test_GeocodeMerger.py
@@ -37,6 +37,12 @@ class TestSmartGeocoder(unittest.IsolatedAsyncioTestCase):
     async def test_3_smart_coding(self):
         gss = SmartLinker(server='OGP')
         await gss.initialise()
+        gss.allow_geometry('connected_tables')
+        gss.run_graph(starting_column='LAD21CD', ending_column='OA21CD', geographic_areas=['Lewisham'], geographic_area_columns=['LAD21NM'])  # the starting and ending columns should end in CD
+
+    async def test_4_smart_coding(self):
+        gss = SmartLinker(server='OGP')
+        await gss.initialise()
         gss.allow_geometry()
         gss.run_graph(starting_column='WD22CD', ending_column='LAD22CD', geographic_areas=['Lewisham', 'Southwark'], geographic_area_columns=['LAD22NM'])  # the starting and ending columns should end in CD
 
@@ -45,7 +51,7 @@ class TestSmartGeocoder(unittest.IsolatedAsyncioTestCase):
         assert gss.fs.chunk_size == 5
         assert codes['table_data'][0]['WD22CD'].nunique() == 42
 
-    def test_4_geo_helper(self):
+    def test_5_geo_helper(self):
         geo_help = GeoHelper(server='OGP')
         print(geo_help.available_geographies())
         geo_keys = geo_help.geography_keys()